Recently, I saw [this post](https://community.netlify.com/t/support-guide-making-the-most-of-netlifys-cdn-cache/127) on the Netlify forums, explaining common problems with large sites causing long deploy times. And if we look at [two](https://app.netlify.com/sites/datenanfragen/deploys/5ec2bb3119ca5700074682b7) [recent](https://app.netlify.com/sites/datenanfragen/deploys/5ec2bbb6ea97040006b95cee) deploys, we see this exact phenomenon on our site: More than 5000 changed files have to be uploaded with each deploy.

That is clearly not a result of the [actual changes](https://github.com/datenanfragen/website/commit/d0928da2ce45d16311fd94211576271b75424297) between these two commits. Luckily, Netlify allows us to download the build results (takes a while though and sometimes needs a couple of tries). If we throw those into a directory diff tool, the problem quickly becomes apparent:

![image](https://user-images.githubusercontent.com/4048809/82943057-7e351c80-9f88-11ea-93ab-f5d7dbeb3000.png)

As I [already noted](https://github.com/datenanfragen/website/blob/2130c1bdcdd3a4f994b2c0c0add3da93a73237d1/layouts/company/single.html#L105) when introducing the related companies feature, the related pages Hugo generates are not deterministic at all. They change with every build, causing every entry in the company database to have a different hash each time.

This PR replaces the Hugo implementation with our own one. Unfortunately, this means the full Hugo build takes a fair bit longer. On my machine, running `hugo -e production --minify` takes:

* without related pages: `2748 ms`
* with the previous implementation: `3713 ms`
* with a one-line intersect: `7822 ms`  
  (`{{ $related := (where .Site.Pages ".Params.categories" "intersect" .Params.categories) | intersect (where .Site.Pages ".Params.relevant-countries" "intersect" (.Param "relevant-countries")) | first 5 }}`)
* with the more complex intersect loop and scoring: `18043 ms`

However, these build times are not really relevant as we have disabled the related companies calculation for local devs builds anyway, so we won't notice them on our machines. I am also fairly confident that the extra 10 seconds will be worth it in CI envs as currently on Netlify the actual build takes about 7 minutes while deploying and processing takes 14 minutes.

---

Note: This PR doesn't address another, analogous problem: We are using absolute URLs in our builds. That means that despite the changes in this PR, deploy previews will still cause huge diffs, at least on the first build. Subsequent builds should be faster.  
I don't believe it is worth it to change that, though. Especially since this doesn't affect production builds.