@baltpeter 2020-05-26
We may also want to [ask them to completely disable post processing](https://community.netlify.com/t/completely-disabling-post-processing-of-deploys/2678) our our site. I'm pretty sure we are already doing all of that in our build pipeline anyway.