@baltpeter 2020-05-30
I've [asked](https://community.netlify.com/t/completely-disabling-post-processing-of-deploys/2678/33?u=benni) for post-processing to be disabled now.

There already seems to be quite an improvement with the second deploy in this PR 'only' taking 13 mins and only 31 vs. 6000 files to upload!

![image](https://user-images.githubusercontent.com/4048809/83333019-03bd1300-a28e-11ea-8652-002a36fbc9fd.png)